# Simplify and optimize migration_cache.Reader